Add files via upload
I added the Store class, which gives users the ability to purchase credits and/or use credits, aswell as transfer credits between players. I added a bank class which is a mock bank to use in place of a real one and it has a withdraw function, a function to add accounts, and a function that lets you check how much money you have. Each user is given $20 and no more can be added during play. The purchase proxy is a class that allows for the bank and the store to interact when purchasing credits. The credits class is used for keeping track of and manipulating credits.